updated main branch with updated version 2.0.3

+++ public/src/js/wallet-system-for-woocommerce-public.js
@@ -116,8 +116,8 @@
 		var amount = $(this).val();
 		var minamount = $(this).data('min');
 		var maxamount = $(this).data('max');
-		minamount = parseInt(minamount);
-		maxamount = parseInt(maxamount);
+		minamount = parseFloat(minamount);
+		maxamount = parseFloat(maxamount);
 		if ( amount <= 0 ) {
 			$('.error').show();
 			$('.error').html(wsfw_public_param.wsfw_amount_error);
